# Conflicts:
#	build.gradle.kts
#	src/main/kotlin/com/milaboratory/mixcr/cli/CommandExtend.kt
#	src/main/kotlin/com/milaboratory/mixcr/cli/CommandFindAlleles.kt
#	src/main/kotlin/com/milaboratory/mixcr/cli/CommandFindShmTrees.kt

+++ src/main/java/com/milaboratory/mixcr/partialassembler/PartialAlignmentsAssemblerParameters.java
@@ -36,7 +36,7 @@ public class PartialAlignmentsAssemblerParameters {
             @JsonProperty("kOffset") int kOffset,
             @JsonProperty("minimalAssembleOverlap") int minimalAssembleOverlap,
             @JsonProperty("minimalNOverlap") int minimalNOverlap,
-            @JsonProperty("minimalAlignmentMergeIdentity") int minimalAlignmentMergeIdentity,
+            @JsonProperty("minimalAlignmentMergeIdentity") float minimalAlignmentMergeIdentity,
             @JsonProperty("mergerParameters") MergerParameters mergerParameters,
             @JsonProperty("maxLeftParts") long maxLeftParts,
             @JsonProperty("maxLeftMatches") int maxLeftMatches) {
